Simple AI Agent/requirements.txt
langchain
wikipedia
langchain-community
langchain_groq
python-dotenv
pydantic
duckduckgo-search